+++ packages/eyeglass/test/test_version.js
@@ -14,7 +14,7 @@ describe("sass version function", function () {
 
  it("should return a module's version", function (done) {
    var options = {
-     root: testutils.fixtureDirectory("basic_modules"),
+     eyeglass: {root: testutils.fixtureDirectory("basic_modules")},
      data: ".version {\nmodule-a: eyeglass-version('module_a')}"
    };
    var expectedOutput = ".version {\n  module-a: 1.0.1; }\n";
@@ -23,7 +23,7 @@ describe("sass version function", function () {
 
  it("should handle asking for the version of an unknown module", function (done) {
    var options = {
-     root: testutils.fixtureDirectory("basic_modules"),
+     eyeglass: {root: testutils.fixtureDirectory("basic_modules")},
      data: ".ok { color: blue} .version {\nmodule-a: eyeglass-version('not_real_module')}"
    };
    var expectedOutput = ".ok {\n  color: blue; }\n";